backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
litellm>=1.55.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
